% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main_functions.R
\name{anno_vars}
\alias{anno_vars}
\title{Annotate variant file from BED files}
\usage{
anno_vars(bed_dir_path, variant_path, cores = 1, write_files = TRUE)
}
\arguments{
\item{bed_dir_path}{A character.}

\item{variant_path}{A data.table.}

\item{cores}{A numeric.}

\item{write_files}{A logical.}
}
\value{
NULL for write_files == FALSE; data.table for write_files == TRUE
}
\description{
Annotate variant file from BED files
}

