% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{annotate_variants}
\alias{annotate_variants}
\title{Annotate a data.table with binary annotations.}
\usage{
annotate_variants(bed_dir_path, variant_path, cores = 1)
}
\arguments{
\item{bed_dir_path}{A character.}

\item{variant_path}{A data.table.}

\item{cores}{A numeric.}
}
\value{
Annotated data.frame.
}
\description{
Annotate a data.table with binary annotations.
}
\seealso{
Other annotation functions with NO intermediate file output: \code{\link{get_anno_col}}
}

