% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main_functions.R
\name{annotate_variants_with_intermediates}
\alias{annotate_variants_with_intermediates}
\title{Write intermediate cols and horizontally concatenate output.}
\usage{
annotate_variants_with_intermediates(bed_dir_path, variant_path, cores = 1)
}
\arguments{
\item{bed_dir_path}{A character.}

\item{variant_path}{A data.table.}

\item{cores}{A numeric.}
}
\value{
NULL.
}
\description{
Write intermediate cols and horizontally concatenate output.
}
\seealso{
Other annotation functions WITH intermediate file output: \code{\link{horizontal_concat_annos}};
  \code{\link{write_anno_col}};
  \code{\link{write_variant_annotations}}
}

