% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{get_anno_col}
\alias{get_anno_col}
\title{Generate binary annotation column.}
\usage{
get_anno_col(bed_file_path, variant_table)
}
\arguments{
\item{bed_file_path}{A character.}

\item{variant_table}{A data.table.}
}
\value{
Annotation column.
}
\description{
Generate binary annotation column.
}
\seealso{
Other annotation functions with NO intermediate file output: \code{\link{annotate_variants}}
}

