% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{get_bed}
\alias{get_bed}
\title{Import BED file.}
\usage{
get_bed(bed_path)
}
\arguments{
\item{bed_path}{A character.}
}
\value{
data.table
}
\description{
Import BED file.
}
\seealso{
Other utility functions: \code{\link{get_file_paths}},
  \code{\link{get_variant_table}}, \code{\link{gzfread}}
}

