% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{get_file_paths}
\alias{get_file_paths}
\title{Get the full paths of files from a directory.}
\usage{
get_file_paths(dir_path)
}
\arguments{
\item{dir_path}{A character.}
}
\value{
character
}
\description{
Get the full paths of files from a directory.
}
\seealso{
Other utility functions: \code{\link{get_bed}},
  \code{\link{get_variant_table}}, \code{\link{gzfread}}
}

