% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{get_variant_table}
\alias{get_variant_table}
\title{Import variant table in pseudo-BED format.}
\usage{
get_variant_table(variant_path)
}
\arguments{
\item{variant_path}{A character.}
}
\value{
data.table
}
\description{
Import variant table in pseudo-BED format.
}
\seealso{
Other utility functions: \code{\link{get_bed}},
  \code{\link{get_file_paths}}, \code{\link{gzfread}}
}

