% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main_functions.R
\name{gzfread}
\alias{gzfread}
\title{Read gzipped file into data.table.}
\usage{
gzfread(path, sep, out_dir = NULL)
}
\arguments{
\item{path}{A character.}

\item{sep}{A character.}
}
\value{
data.table
}
\description{
Read gzipped file into data.table.
}
\seealso{
Other utility functions: \code{\link{get_bed}};
  \code{\link{get_file_paths}};
  \code{\link{get_variant_table}}
}

