% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main_functions.R
\name{horizontal_concat_annos}
\alias{horizontal_concat_annos}
\title{Horizontally concatenate intermediate annotation columns.}
\usage{
horizontal_concat_annos(variant_path, anno_col_dir_path)
}
\arguments{
\item{variant_path}{A character.}

\item{anno_col_dir_path}{A character.}
}
\value{
NULL.
}
\description{
Horizontally concatenate intermediate annotation columns.
}
\seealso{
Other annotation functions WITH intermediate file output: \code{\link{annotate_variants_with_intermediates}};
  \code{\link{write_anno_col}};
  \code{\link{write_variant_annotations}}
}

