% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{write_anno_col}
\alias{write_anno_col}
\title{Write intermediate annotation column.}
\usage{
write_anno_col(bed_file_path, variant_table, anno_path)
}
\arguments{
\item{bed_file_path}{A character.}

\item{variant_table}{A data.table.}

\item{anno_path}{A character.}
}
\value{
NULL.
}
\description{
Write intermediate annotation column.
}
\seealso{
Other annotation functions WITH intermediate file output: \code{\link{annotate_variants_with_intermediates}},
  \code{\link{horizontal_concat_annos}},
  \code{\link{write_variant_annotations}}
}

