% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main_functions.R
\name{write_variant_annotations}
\alias{write_variant_annotations}
\title{Write intermediate annotation columns from dir.}
\usage{
write_variant_annotations(bed_dir_path, variant_path, cores = 1)
}
\arguments{
\item{bed_dir_path}{A character.}

\item{variant_path}{A data.table.}

\item{cores}{A numeric.}
}
\value{
NULL.
}
\description{
Write intermediate annotation columns from dir.
}
\seealso{
Other annotation functions WITH intermediate file output: \code{\link{annotate_variants_with_intermediates}};
  \code{\link{horizontal_concat_annos}};
  \code{\link{write_anno_col}}
}

